✨ Add unit tests for JobService, SolverService, and CatalogService with fixtures
